// src/lib/types/report.ts
// src/lib/types/report.ts
export interface Table31ReportData {
  discipline: string;
  fac_nip: string;
  fac_name: string;
  specialty_field1: string;
  specialty_field2: string;
  highest_degree: string;
  highest_degree_year: number;
  normal_professional_responsibilities: string;
  fac_time: number;
  sa: number;
  pa: number;
  sp: number;
  ip: number;
  a: number;
}